Add files via upload
Maintaining the temperature in the smart office by having a controller connected to the home router. The controller is connected to the display, which shows the temperature for everyone without any complications. And also have a so-called temperature monitoring unit whose function is similar to the function of the display screen, but have added adaptation to it and conditions for example, if there is an increase in the temperature of more than 30, the air conditioner is turned on  automatically, and in case the opposite becomes any Below 20, when the temperature is low, the heaters are turned on automatically. also made an application for a laptop and tablet. We have made a link that you can access and control all devices with the addition of a password and name (to verify and prevent attacks) And  have added a build supervisor and he has been given the controller by the authenticator app and we have added the password to him.